Settings fragment and Add User
Added a very basic settings fragment so that button in the nav drawer will actually do something, it only links to Add User which doesn't actually do anything. Ideally, it should send off a call to the API to add the new user. It should also use a dropdown instead of text field for the room, which could maybe be pulled from the backend when the page is loaded.